orientation-event/DeviceMotion/t021.html

<!DOCTYPE html>
<html>

  <head>
	<title>User agents implementing this specification must provide a new DOM event, named devicemotion:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
  <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>User agents implementing this specification must provide a new DOM event, named devicemotion:</h1>
    <div id="log"></div>
	
	<script type="text/javascript">
		t = null;
		at = async_test();
		
		window.addEventListener('devicemotion', function(e) {t = true}, false);
		setTimeout("testing()", 200);
		
		function testing(){
			at.step(function() {assert_true(t)}); 
			at.done();
		}
	</script>
	
  </body>
  
</html>
